renamed the examples to diffdrivebot and bicycledrivebot_carlike
