--- example_11/README.md
@@ -1,6 +1,0 @@
-# ros2_control_demo_example_11
-
-   *CarlikeBot*, or ''Carlike Mobile Robot'', is a simple mobile base with bicycle drive.
-   The robot has two wheels in the front that steer the robot and two wheels in the back that power the robot forward and backwards. However, since each pair of wheels (steering and traction) are controlled by one interface, a bicycle steering model is used.
-
-Find the documentation in [doc/userdoc.rst](doc/userdoc.rst) or on [control.ros.org](https://control.ros.org/master/doc/ros2_control_demos/example_11/doc/userdoc.html).
