--- example_11/description/ros2_control/carlikebot.ros2_control.xacro
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<robot xmlns:xacro="http://www.ros.org/wiki/xacro">
-
-  <xacro:macro name="carlikebot_ros2_control" params="name prefix">
-
-    <ros2_control name="${name}" type="system">
-      <hardware>
-        <plugin>ros2_control_demo_example_11/CarlikeBotSystemHardware</plugin>
-        <param name="example_param_hw_start_duration_sec">0</param>
-        <param name="example_param_hw_stop_duration_sec">3.0</param>
-      </hardware>
-      <joint name="${prefix}virtual_front_wheel_joint">
-        <command_interface name="position"/>
-        <state_interface name="position"/>
-      </joint>
-      <joint name="${prefix}virtual_rear_wheel_joint">
-        <command_interface name="velocity"/>
-        <state_interface name="velocity"/>
-        <state_interface name="position"/>
-      </joint>
-    </ros2_control>
-
-  </xacro:macro>
-
-</robot>
