--- example_2/test/test_urdf_xacro.py
@@ -1,77 +1,0 @@
-# Copyright (c) 2022 FZI Forschungszentrum Informatik
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#    * Redistributions of source code must retain the above copyright
-#      notice, this list of conditions and the following disclaimer.
-#
-#    * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-#
-#    * Neither the name of the {copyright_holder} nor the names of its
-#      contributors may be used to endorse or promote products derived from
-#      this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-#
-# Author: Lukas Sackewitz
-
-import os
-import shutil
-import subprocess
-import tempfile
-
-from ament_index_python.packages import get_package_share_directory
-
-
-def test_urdf_xacro():
-    # General Arguments
-    description_package = "ros2_control_demo_example_2"
-    description_file = "diffbot.urdf.xacro"
-
-    description_file_path = os.path.join(
-        get_package_share_directory(description_package), "urdf", description_file
-    )
-
-    (_, tmp_urdf_output_file) = tempfile.mkstemp(suffix=".urdf")
-
-    # Compose `xacro` and `check_urdf` command
-    xacro_command = (
-        f"{shutil.which('xacro')}" f" {description_file_path}" f" > {tmp_urdf_output_file}"
-    )
-    check_urdf_command = f"{shutil.which('check_urdf')} {tmp_urdf_output_file}"
-
-    # Try to call processes but finally remove the temp file
-    try:
-        xacro_process = subprocess.run(
-            xacro_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
-        )
-
-        assert xacro_process.returncode == 0, " --- XACRO command failed ---"
-
-        check_urdf_process = subprocess.run(
-            check_urdf_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
-        )
-
-        assert (
-            check_urdf_process.returncode == 0
-        ), "\n --- URDF check failed! --- \nYour xacro does not unfold into a proper urdf robot description. Please check your xacro file."
-
-    finally:
-        os.remove(tmp_urdf_output_file)
-
-
-if __name__ == "__main__":
-    test_urdf_xacro()
