+++ example_bicycledrivebot_carlike/README.md
@@ -1,0 +1,6 @@
+# ros2_control_demo_bicycledrivebot_carlike
+
+Changes:
+- renamed ros2_control_demo_example_11 to ros2_control_demo_bicycledrivebot_carlike
+
+Find the documentation of ros2_control_demo_example_11 on [control.ros.org](https://control.ros.org/master/doc/ros2_control_demos/example_11/doc/userdoc.html).
