+++ example_bicycledrivebot_carlike/hardware/include/ros2_control_demo_bicycledrivebot_carlike/carlikebot_system.hpp
@@ -1,0 +1,93 @@
+// Copyright 2021 ros2_control Development Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef ROS2_CONTROL_DEMO_BICYCLEDRIVEBOT_CARLIKE__CARLIKEBOT_SYSTEM_HPP_
+#define ROS2_CONTROL_DEMO_BICYCLEDRIVEBOT_CARLIKE__CARLIKEBOT_SYSTEM_HPP_
+
+#include <map>
+#include <memory>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "hardware_interface/handle.hpp"
+#include "hardware_interface/hardware_info.hpp"
+#include "hardware_interface/system_interface.hpp"
+#include "hardware_interface/types/hardware_interface_return_values.hpp"
+#include "rclcpp/clock.hpp"
+#include "rclcpp/duration.hpp"
+#include "rclcpp/macros.hpp"
+#include "rclcpp/time.hpp"
+#include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
+#include "rclcpp_lifecycle/state.hpp"
+
+namespace ros2_control_demo_bicycledrivebot_carlike
+{
+struct JointValue
+{
+  double position{0.0};
+  double velocity{0.0};
+  double effort{0.0};
+};
+
+struct Joint
+{
+  explicit Joint(const std::string & name) : joint_name(name)
+  {
+    state = JointValue();
+    command = JointValue();
+  }
+
+  Joint() = default;
+
+  std::string joint_name;
+  JointValue state;
+  JointValue command;
+};
+class CarlikeBotSystemHardware : public hardware_interface::SystemInterface
+{
+public:
+  RCLCPP_SHARED_PTR_DEFINITIONS(CarlikeBotSystemHardware);
+
+  hardware_interface::CallbackReturn on_init(
+    const hardware_interface::HardwareInfo & info) override;
+
+  hardware_interface::CallbackReturn on_configure(
+    const rclcpp_lifecycle::State & previous_state) override;
+
+  hardware_interface::CallbackReturn on_activate(
+    const rclcpp_lifecycle::State & previous_state) override;
+
+  hardware_interface::CallbackReturn on_deactivate(
+    const rclcpp_lifecycle::State & previous_state) override;
+
+  hardware_interface::return_type read(
+    const rclcpp::Time & time, const rclcpp::Duration & period) override;
+
+  hardware_interface::return_type write(
+    const rclcpp::Time & time, const rclcpp::Duration & period) override;
+
+private:
+  // Parameters for the CarlikeBot simulation
+  double hw_start_sec_;
+  double hw_stop_sec_;
+
+  // joint names
+  std::string steering_joint_;
+  std::string traction_joint_;
+};
+
+}  // namespace ros2_control_demo_bicycledrivebot_carlike
+
+#endif  // ROS2_CONTROL_DEMO_BICYCLEDRIVEBOT_CARLIKE__CARLIKEBOT_SYSTEM_HPP_
