+++ example_diffdrivebot/README.md
@@ -1,0 +1,6 @@
+# ros2_control_demo_diffdrivebot
+
+Changes:
+- renamed ros2_control_demo_example_2 to ros2_control_demo_diffdrivebot 
+
+Find the documentation of ros2_control_demo_example_2 on [control.ros.org](https://control.ros.org/master/doc/ros2_control_demos/example_2/doc/userdoc.html).
